// frontend/src/modules/WorkoutsQuiz/WorkoutQuiz.enums.ts
export enum SelectedQuestionQuantity {
  DefaultQuestionQuantity = '',
  FifteenQuestions = '15',
  TwentyQuestions = '20',
  TwentyFiveQuestions = '25',
  ThirtyQuestions = '30',
  ThirtyFiveQuestions = '35',
  FortyQuestions = '40',
}

export enum QuizTimer {
  OneMinute = 60,
  MinuteAndAHalf = 90,
  TwoMinutes = 120,
  TwoMinutesAndHalf = 150,
  ThreeMinutes = 180,
  ThreeMinutesAndHalf = 210,
  FourMinutes = 240,
}